Fixed errorspam

+++ Tiles/Furniture/Razewood/RazewoodLamp.cs
@@ -97,7 +97,7 @@ namespace AAMod.Tiles.Furniture.Razewood
 		}
         public override void PostDraw(int i, int j, SpriteBatch spriteBatch)
         {
-            ulong randSeed = Main.TileFrameSeed ^ (ulong)((long)j << 32 | (long)((ulong)i));
+            ulong randSeed = Main.TileFrameSeed ^ (ulong)(j<< 32 | (long)((ulong)i));
             Color color = new Color(100, 100, 100, 0);
             int frameX = Main.tile[i, j].frameX;
             int frameY = Main.tile[i, j].frameY;
